Improvement approved and merged. 

Improvement to Custom Actions. Only deleted custom actions related to…

+++ Sherpa.Library/SiteHierarchy/CustomActionsManager.cs
@@ -11,8 +11,14 @@ namespace Sherpa.Library.SiteHierarchy
     {
         private static readonly ILog Log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
 
-        public void SetUpCustomActions(ClientContext context, List<ShCustomAction> customActions)
+        public void SetUpCustomActions(ClientContext context, string CustomActionsPrefix, List<ShCustomAction> customActions)
         {
+            if (CustomActionsPrefix != null || CustomActionsPrefix.Equals(String.Empty))
+            {
+                Log.Info("You need to set the property 'CustomActionsPrefix' which will be used for the Custom Action Name.");
+                return; 
+            }
+
             Log.Info("Adding custom actions");
             Site site = context.Site;
             context.Load(site);
@@ -22,7 +28,10 @@ namespace Sherpa.Library.SiteHierarchy
             for (var i = site.UserCustomActions.Count - 1; i >= 0; i--)
             {
                 var customAction = site.UserCustomActions[i];
-                customAction.DeleteObject();
+                if (customAction.Name.StartsWith(CustomActionsPrefix))
+                {
+                    customAction.DeleteObject();
+                }
             }
 
             if (context.HasPendingRequest)
@@ -45,15 +54,13 @@ namespace Sherpa.Library.SiteHierarchy
                 userCustomAction.Sequence = customAction.Sequence;
                 userCustomAction.ScriptSrc = customAction.ScriptSrc;
                 userCustomAction.ScriptBlock = customAction.ScriptBlock;
-
+                userCustomAction.Name = CustomActionsPrefix + "_" + customAction.ScriptSrc.Split('/')[customAction.ScriptSrc.Split('/').Length - 1].Replace(".", "");
                 userCustomAction.Description = customAction.Description;
-                userCustomAction.RegistrationId = customAction.RegistrationId;
                 userCustomAction.RegistrationType = customAction.RegistrationType;
                 userCustomAction.Title = customAction.Title;
-                userCustomAction.Name = customAction.Name;
                 userCustomAction.ImageUrl = customAction.ImageUrl;
                 userCustomAction.Group = customAction.Group;
-
+                
                 try
                 {
                     userCustomAction.Update();
